+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 ## [Unreleased]
 
 ### Fixes
-
 - Bump dependencies versions and fix mocking of redux-sagas/effects
+- Changed redirecting user after reverting to the original file with data source
 - Prompt user when leave route with unsaved changes
 
 ### Changes
-
 - README.md and documentation update
 
 ## 1.6.0 (August 26, 2020) - Release